logs_metrics/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "logs-metrics-service"
version = "0.1.0"
description = "Logging and Metrics Service for IPL Tweet Generator"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "prometheus-client>=0.20.0",
    "uvicorn>=0.30.0",
    "pydantic>=2.7.0",
    "python-dotenv>=1.1.0",
    "requests>=2.31.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.7.0",
    "ruff>=0.0.284"
]

[tool.hatch.build.targets.wheel]
packages = ["."]
